+++ CHANGELOG.md
@@ -9,11 +9,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 ### Added
 
 * Configuration to always use the fast tool instead of the slowest efficient tool.
+  * Accessible via [ModMenu](https://github.com/TerraformersMC/ModMenu).
+  * `/switcheroo alwaysFastest`
 * Configuration for the minimum durability.
+  * Accessible via [ModMenu](https://github.com/TerraformersMC/ModMenu).
+  * `/switcheroo minDurability`
 
 ### Changed
 
-* Saves items when they have 5 durability instead of 1.
+* Don't consider items when they have 5 durability instead of 1.
 
 ### Deprecated